backend/requirements.txt
git+git://github.com/rubengrill/django-nopassword.git@rest-auth#egg=django-nopassword
#django-nopassword==5.0.0
cloudinary==1.18.2
coreapi==2.3.3
dj-database-url==0.5.0
Django==2.2.7
gunicorn==19.9.0
django-admin-sortable==2.1.18
django-cors-headers==3.1.1
django-custom-user==0.7
django-debug-toolbar==2.0
django-nose==1.4.6
django-rest-auth==0.9.5
djangorestframework==3.10.3
flake8-isort==3.0.1
flake8-quotes==2.1.1
flake8==3.7.9
pep8-naming==0.9.0
psycopg2-binary==2.8.4
mailjet-rest==1.3.3
coverage==4.5.4
raven==6.10.0
rednose==1.3.0
whitenoise==4.1.4
user-agents==2.0
requests==2.22.0
ptvsd == 4.1.4
stripe==2.42.0
django_extensions==2.2.6
PyJWT==1.7.1
braintree==3.58.0
hubspot3==3.2.46
short_url==1.2.2
python-dateutil==2.8.1
django-translations==1.1.2
django-nested-inline==0.4.2
django-admin-interface==0.13.5